requirements.txt
pandas
numpy
requests
python-dotenv
yfinance
scikit-learn
xgboost
matplotlib
seaborn
joblib
streamlit
plotly
vaderSentiment
